mostly broken! rules to static constexpr struct

+++ src/one2IIImain.cpp
@@ -43,7 +43,8 @@ int main()
 	uint8_t digits = static_cast<uint8_t>(numberSequence.size());
 	for(uint8_t parts = digits; parts > 0; --parts)
 	{
-		auto partitions = integerPartitions(digits, parts, ruleBook.getMinInputSize(), ruleBook.getMaxInputSize());
+		auto partitions = integerPartitions(digits, parts,
+								(uint8_t)ruleBook.getMinInputSize(), (uint8_t)ruleBook.getMaxInputSize());
 		for(auto& partition : partitions)
 		{
 			do
